+++ src/extensions/react-flow/nodes/NodeContent.tsx
@@ -124,6 +124,11 @@ export interface NodeContentProps<NODE_DATA, NODE_CONTENT_PROPS = any> extends N
      * If set then it will be always overwritten internally.
      */
     selected?: boolean;
+    /**
+     * Allow react flow wheel events, e.g. for zooming using the mouse wheel over a node.
+     * If this is allowed scrolling inside a node is not possible.
+     */
+    letPassWheelEvents?: boolean;
 }
 
 interface MemoHandlerProps extends HandleProps {
@@ -226,6 +231,7 @@ const MemoHandler = React.memo(
      sourcePosition = Position.Right,
      isConnectable = true,
      selected,
+     letPassWheelEvents = false,
      // businessData is just being ignored
      businessData,
      // other props for DOM element
@@ -279,7 +285,8 @@ const MemoHandler = React.memo(
                      ` ${eccgui}-graphviz__node--minimal-${minimalShape}` +
                      (!!highlightedState ? " " + gethighlightedStateClasses(highlightedState, `${eccgui}-graphviz__node`) : "") +
                      (animated ? ` ${eccgui}-graphviz__node--animated` : "") +
-                     (showUnconnectableHandles === false ? ` ${eccgui}-graphviz__node--hidehandles` : "")
+                     (showUnconnectableHandles === false ? ` ${eccgui}-graphviz__node--hidehandles` : "") +
+                     (letPassWheelEvents === false ? ` nowheel` : "")
                  }
              >
                  <header className={`${eccgui}-graphviz__node__header`}>